+++ CHANGELOG.md
@@ -12,6 +12,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - changing a size of `Konva.Stage` will update it in async way (via `batchDraw`).
 - `shadowOffset` respect pixel ratio now
 
+### Fixed
+- Fixed bug when `Konva.Tag` width was not changing its width dynamically
+
 ## [1.2.2][2016-09-15]
 
 ### Fixed